tta_backend/requirements.txt
Django
djangorestframework
djangorestframework-simplejwt
mysqlclient
python-decouple
